fix(cli): Fix plugin loading with rewrite of loading logic (#9171)
**Problem**
Although the cli plugin loading behaviour was functional for our
storybook and data-migrate needs, some flaws where highlighted when it
came to integrate third party packages. The loading logic was somewhat
scatterbrained and had no tests.

**Changes**
1. Reworks the plugin loading logic and, hopefully, makes the logic
easier to follow.
2. Adds tests for the loading logic. This will help catch unintentional
changes in behaviour in future updates.
3. I removed the `version` option from the toml config for experimental
cli plugins. If users wish to enforce a specific version they should
install such a version in the standard way they would for any third
party package - using the package.json and yarn.

**Notes**
1. This is still considered an "experimental" feature so we can
potentially be more loose - which is why I felt a larger rewrite was
advantageous in this case.