feat(mailer): Resend handler (#9175)
**Description**
This is intended to be a mail handler which specifically uses the
[resend node sdk](https://github.com/resendlabs/resend-node/tree/main).

**Changes**
1. Adds a new resend mail handler which uses the resend node sdk to send
mail.
2. Removes an unneeded `uuid` dependency from the in-memory mail
handler. I noticed this when copying over the handler as a template for
this new one.

**Notes**
1. I had to add a transformation of the attachment contents when they
were passed as a string. If I did not transform them into a buffer then
the attachment content would be malformed during my testing.
2. I have tested this manually but for the moment there is no specific
testing suite/strategy in place to test these mail handlers/renderers.