feat: add experimental docker setup command (#9024)
Adds an experimental setup command to add a Dockerfile and compose
files:

```
yarn rw exp setup-docker
```

The Dockerfile is near enough complete that we should release it
experimentally so that we can get feedback from the community. I've
successfully deployed it to Coherence and Fly. As far as I can tell,
Render has half support. The api side runs just fine, but the web side
has a static runtime, and I'm not sure how to get it to take the files
from the web build stage.

There's a known sticking point here: seeding and data migrations. I
understand the problem clearly now. On all providers, seed and data
migrations run using the image built from targeting api serve. There's
only production dependencies in that image—`yarn rw exec` won't work for
a number of reasons, but the main one is that `@redwoodjs/internal`
isn't around. It's looking like we'll have to move `yarn rw exec` into
it's own package and that we'll have to finish
https://github.com/redwoodjs/redwood/pull/8572.

The docs are there but are very much a WIP.