fix(crwa): Explicit check for possible null value in `entry.client.tsx` (#9251)
**Problem**
Fixes #9059 which is a type error raised when you have strict type
checking enabled. The `getElementById` call can return null as the [MDN
docs](https://developer.mozilla.org/en-US/docs/Web/API/Document/getElementById#return_value)
mention.

**Changes**
Adds a basic null check taken from the solution given in #9059. 

**Considerations**
This adds some boilerplate code to a user facing file but I think it's a
reasonable addition. It's readable and clear what the code is doing
without the need for further explanation or technical knowledge. In the
case where this causes a runtime error then the message is clearer:
![Screenshot from 2023-10-02
23-16-38](https://github.com/redwoodjs/redwood/assets/56300765/ca6d51bd-2c28-438e-87d3-7411e54721ec)

I haven't immediately been able to think of some hidden solution to this
that isn't exposed to the user in some boilerplate code.