chore(api-server): improve tests (#9325)
This PR improves the tests for the `@redwoodjs/api-server` package. In
doing this, I deliberately didn't edit any sources files (save one edit
to export an object for testing).

My goal was basically to test everything I could. I'm going to be
deduplicating code across the framework (a lot of the code in the
api-server package was duplicated to avoid breaking anything while
iterating on experimental features), and since users can invoke
api-server code in a few different ways (`npx @redwoodjs/api-server`,
`yarn rw serve`, and importing handlers from `@redwoodjs/api-server`), I
don't want to accidentally break anything anywhere. Improving test
coverage also revealed some things that need to be fixed.

A few high-level notes. Overall, I reduced our reliance on mocking (in
many tests, we weren't using fastify at all). Most tests covered
`configureFastify` and while a few checked to see that routes were
registered correctly, none of them actually started the api-server, or
used any of fastify's testing facilities like `inject` to check that the
right file was sent back, etc. (Now they do!) Lastly, while most of the
tests I added assume fastify, the `dist.test.ts` isn't and is more like
a black-box test.

Re the "revealed some things that need to be fixed", in no particular
order:

- `apiUrl` and `apiHost` can be a bit confusing. Even the previous
`withApiProxy.test.ts` got them wrong

While this test checked if the options were passed correctly, it mixed
up the `apiUrl` for the `apiHost`. `apiHost` is the upstream (and should
be a URL).

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/__tests__/withApiProxy.test.ts#L51-L54

@Josh-Walker-GM alerted me to this some time ago, but I didn't
understand the nuance. Now that I do, these are just poorly named and we
should try to fix them.

- Everything in web dist is served

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/plugins/withWebServer.ts#L50-L53

We may want to revisit this. I'm sure most of it needs to be served, but
it seems like a poor default when it comes to security. Also, it just
results in routes being served that don't work. E.g., all prerendered
routes are available at their `.html` suffix, but they error out as soon
as they're rendered.

- The lambda adapter seems a little too simple

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/requestHandlers/awsLambdaFastify.ts#L11-L27

There's a [core fastify plugin for
lambdas](https://github.com/fastify/aws-lambda-fastify), and it's much
more complicated than our implementation. Should we just use that?
What's the difference between our implementation and theirs?

- `setLambdaFunctions` should ignore functions that don't export a
handler.

  It probably doesn't do any harm but it registers them as `undefined`:

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/plugins/lambdaLoader.ts#L32-L40

- The lambda adapter supports callback syntax

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/requestHandlers/awsLambdaFastify.ts#L71-L89

  I don't think we need to support this anymore.

- The CLI handlers don't have help and don't error out on unknown args

  It's not uncommon to use this CLI standalone so we should polish it.

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/index.ts#L26-L28

  ```
  $ yarn rw-server --help
  Options:
--help Show help [boolean]
--version Show version number [boolean]
  ```

  ```
  $ yarn rw-server --foo --bar --baz
  Starting API and Web Servers...
  ```

- You can't import `@redwoodjs/api-server` outside a Redwood project

This one isn't relevant to users, but it made testing the built package
more complicated than it needed to be. The issue is that `getConfig` is
called as soon as the file is imported. And if you're not in a redwood
project at that point, you're out of luck:

https://github.com/redwoodjs/redwood/blob/daaa1998837bdb6eaa42d9160292e781fadb3dc8/packages/api-server/src/cliHandlers.ts#L24

---------

Co-authored-by: Josh GM Walker <56300765+Josh-Walker-GM@users.noreply.github.com>