fix(babel): Fix opentelemetry api wrapping and allow it to be disabled (#9298)
**Problem**
Some projects fail to build when opentelemetry is enabled. This is
because the babel plugin which automatically wraps the api side
functions failed to handle some syntax cases. In those cases the
transpiled syntax would be invalid javascript.

**Changes**
1. Supports default values in function arguments.
2. Supports some `AssignmentPattern` function arguments.
3. Add test fixtures. We can now see if the transpiled code is valid and
as we expect.
4. Allow the api wrapping to be disabled via a `wrapApi` toml option
5. Expose the `wrapApi` toml option during opentelemetry setup.

**Notes**
There are still some syntax cases which aren't supported in the babel
plugin. In this case we now bail out and do not modify the users source
code rather than transpile into nonsense syntax.

---------

Co-authored-by: Dominic Saadi <dominiceliassaadi@gmail.com>