feat(cli): Setup command for mailer (#9335)
**Description**
Mailer will be available from a setup command: `yarn rw setup mailer`.
This PR adds that command.

Like all good PRs this also does a second thing. This adds functionality
to the mailer which enables it to conditionally use an in-memory handler
for test and the studio handler for development if those dependencies
are available. This is essentially adding automatic defaults but
respecting if the dependencies are available.

**Changes**
1. Adds a setup command: `yarn rw setup mailer`
2. Implements fallback/default dev & test handlers for when the user has
not set specific configuration.

**Notes**
1. I'm not overly thrilled with the typing/implementation of these
automatic/fallback test/dev handlers but we can improve that going
forward.

**Future work**
1. We should improve the logging output.
2. I NEED to write the docs now.

---------

Co-authored-by: David Thyresson <dthyresson@gmail.com>