fix(server-file): prefix `MODULE_NOT_FOUND` with `ERR_` (#9372)
If you try to `await import` a module and it can't be found, it turns
out Node throws a different error than if you try to `require` a module
and it can't be found:

- `require(mod)` -> `MODULE_NOT_FOUND`
- `await import(mod)` -> `ERR_MODULE_NOT_FOUND`

We used to tranpsile `await import`s to `require`. When we stopped doing
that, we forgot to change the error codes we were checking.