chore: Relocated tsbuildinfo out of dist folder (#9390)
**Problem**
The current tsconfig entry for the tsbuildinfo file puts the file into
the dist folder. This ultimately results in it being included in the
shipped packages on npm. These files are large and in some cases are the
bulk of the content in the package. For example in the
`@redwoodjs/project-config` package it makes up 75% of the package size.

From what I can read online, and my own understanding of the
functionality of that file, there is no need for this content to be
shipped with the built code on npm.

I quickly checked on my local machine and this change does not appear to
have impacted the build time.

**Change**
1. Delete the explicit entry which will place it in a default location
which is fine. The file is already ignored by the existing .gitignore.
2. Remove the code in the release tooling that removes these files. I
don't think it was fully functional right now.

**Notes**
I marked this as a chore because it doesn't influence the code just the
build step. It does change the bytes that are shipped to users so maybe
it's not a chore but I am not that concerned with the label right now.