fix(v6.4.0 RC): conditionally remove `SSELink` from the bundle (#9401)
Follow up to #9009 and #9205.

#9009 increased the bundle size of Redwood apps:

<table>
<tr><th>stable (v6.3.3)</th><th>v6.4.0 RC</th><th>% increase</th></tr>
<tr><td>

- Rendered: 1.03MB (100.00%)
- Gzip: 338.99KB
- Brotli: 290.87KB
    
</td><td>

- Rendered: 1.46MB (100.00%)
- Gzip: 461.74KB
- Brotli: 394.16KB

</td><td>

- Rendered: 41.75%
- Gzip: 36.21%
- Brotli: 35.51%

</td></tr>
</table>

I'm using
[rollup-plugin-visualizer](https://www.npmjs.com/package/rollup-plugin-visualizer)
on a newly-created Redwood app to get these numbers and treemaps. Here's
links to the full treemaps:
-
[stable](https://github.com/redwoodjs/redwood/assets/32992335/41cf4204-10de-4b69-a07a-7d6b980ead31)
- [v6.4.0
RC](https://github.com/redwoodjs/redwood/assets/32992335/f1a6a5bf-a260-45cb-8c97-f68b4fe0730d)

We don't have a history of bundle size metrics, so it's hard to tell
what the magnitude of this change is. But obviously in general, an
increase isn't good, especially if it's for a feature that isn't used
across the board.

One solution was mentioned in this thread:
https://github.com/redwoodjs/redwood/pull/9009#discussion_r1299941377.
It was to have users import the link, that way if they didn't, Vite
wouldn't bundle it. The pushback there was more config for the user.

I paired with @Josh-Walker-GM on a solution that doesn't involve any
tradeoffs, except that it was work for us: we use Vite to remove the
file from the bundle if realtime isn't configured. No bundle size
increase if you're not using realtime, no config if you are!

Note that we need to watch out for
https://github.com/redwoodjs/redwood/pull/9342 since it removes the
option from the toml file that this solution uses.

This was the main issue blocking v6.4.0, so now that it's resolved,
we're ready to ship next week.

**Update**

This PR was failing the serve smoke test consistently. At first I
thought it was because I didn't add the same logic to our webpack
config, so I went ahead and did that but it was still failing. After a
few dead ends, it turns out the logic I naively copied from the router
to conditionally render the splash page has had a bug all along:


https://github.com/redwoodjs/redwood/blob/699723910904d243ab153833b2d1a8e7494fcf40/packages/router/src/router.tsx#L120

You can see it easily by building and serving a Redwood app that doesn't
have a home page:

<img width="1186" alt="image"
src="https://github.com/redwoodjs/redwood/assets/32992335/b4ee1369-af45-4887-aacf-3217212bdc68">

This is because this line is transpiled into:

```js
var _splashPage = requireSplashPage();
// ...
if (shouldShowSplash && typeof _splashPage.SplashPage !== "undefined") {
```

And you can't get away with `undefined.SplashPage`. The reason we didn't
see it before is that hardly any user run `yarn rw serve` without a home
page.

Fixed by changing the module from a default `null` export to an empty
object export:
https://github.com/redwoodjs/redwood/pull/9401/commits/591ee5afbf9c0f3c153e82059daffa0d6acc9400.