chore(ci): Fix conditional workflow jobs (#9417)
**Problem**
We have been a number of workflow jobs that we intend to be conditional
based on what changes actually occurred in the PR. Recently this has
caused a small issue where docs PRs are being held up by CI expecting a
number of required workflow jobs to execute but they never will because
we don't need them to.

**Changes**
1. Centralises the docs/rsc/ssr change detection to a single action
which provides multiple outputs
2. Renaming to include `-skip` job suffixes for workflow jobs which are
intended to run simply to pass the required check status.
3. Refactoring - e.g. reordering the workflow jobs

**Notes**
1. You may examine our working on this
[here](https://github.com/orgtoar/redwood-test/pull/320).
2. There are example PRs for different changes
[here](https://github.com/orgtoar/redwood-test/actions/runs/6868994127/job/18680823427?pr=321),
[here](https://github.com/orgtoar/redwood-test/actions/runs/6869039658?pr=322),
[here](https://github.com/orgtoar/redwood-test/actions/runs/6869087419?pr=323)
and
[here](https://github.com/orgtoar/redwood-test/actions/runs/6869099868?pr=324).
This should show the conditional jobs working as expected.

Thanks @jtoar for helping me understand this better!