docs(docker): add troubleshooting section for python (#9548)
I'm working with @KrisCoulson on the experimental Dockerfile. His
project needs Python, and I'm guessing others will too. I'm still torn
on adding it to the Dockerfile since it's not a small addition, so till
I have more clarity on that, the docs should at least highlight it
because it's quite simple to fix.