fix(crwa): pin Node.js version in `.nvmrc` (#9614)
@Tobbe and I reproduced Netlify deploys failing on Node.js with the
following Prisma error (only during run, not build, curiously):

```
Prisma Client could not locate the Query Engine for runtime "rhel-openssl-3.0.x".This happened because Prisma Client was generated for "debian-openssl-1.1.x", but the actual deployment required "rhel-openssl-3.0.x"
```

Netlify was deploying with Node.js 20, even if we set the `NODE_VERSION`
env var to `18` in Netlify's deploy settings. It seemed like the
`.nvmrc` file (which specified any lts version: `lts/*`) was taking
precedence over the NODE_VERSION env var.

This PR pins the Node.js lts version to 18 in the .nvmrc. We're not sure
if we even need to ship with this file but this seemed like the safer
change for now.

Update: @Tobbe and I tried removing the `.nvmrc` file in a test deploy.
Netlify defaulted to Node.js 16 which also isn't what we want, so it
should stay.