chore(cli): make fs modules used in the CLI consistent (#9746)
I was sizing up the CLI code as part of ESM work. Since I was looking at
it wholesale, @Josh-Walker-GM and I were exchanging notes on what we'd
like to improve about it. While a lot of what we talked about was out of
scope, there were a few things that are low hanging fruit, one of which
is this one, which amounted to a simple search and replace. We just
figured if we're using `fs-extra` half the time in the CLI, we may as
well use it everywhere to remove the overhead of why it'd be different
in one place vs another.