chore(crwa): add e2e tests for create-redwood-app (#9783)
Similar to https://github.com/redwoodjs/redwood/pull/9766, I was working
on ESM and started with CRWA since it's the easiest package, and there
really weren't any tests for it. (I think other tests, like the cypress
e2e test, use it, so I think we'd catch if it totally broke, but it's
not trivial to run that one locally for quick feedback.) This PR adds a
couple of e2e tests.

To run them locally:

```bash
yarn build

cd ./packages/create-redwood-app
yarn build:pack

PROJECT_PATH=$(yarn set-up-test-project)

PROJECT_PATH=$PROJECT_PATH yarn test:e2e
PROJECT_PATH=$PROJECT_PATH ./tests/e2e_prompts.sh
```

Some of the CRWA e2e tests use
[expect](https://core.tcl-lang.org/expect/index). I tried testing the
prompts using `zx`, Node's `child_process`, etc and writing to
`process.stdin`, but it never worked consistently and when it did it was
hacky. Expect more or less just worked. Those tests are limited to
Ubuntu, but better than the nothing we had before.

In making the ts-to-js script ESM, I was running it and it seems like
the JS template was a little out of date. So updated that as well.

A couple of other notes

- i got rid of the `.babelrc.js` file which was only there for jest. all
the tests are ESM now

- since i added a new script (setUpTestProject), i made a new scripts
directory and added tsToJS to it too

- one e2e test i don't have yet is for when the node version is lower or
higher. i couldn't seem to set it up with expect. will keep thinking
about how to test that one

- it's likely that the expect tests could be consolidated or just be
better in general

- every invocation of `yarn create-redwood-app` seems to be printing
some escape codes to `stderr`:

   ```js
   const p = await $`yarn create-redwood-app --help`
   expect(p.stderr).toMatchInlineSnapshot(`"[?25l[?25h"`)
  ```

- `yarn create-redwood-app --version` prints the banner and some escape
codes; both seem unnecessary:

   ```js
   const p = await $`yarn create-redwood-app --version`

   expect(p.exitCode).toEqual(0)
   expect(p.stdout).toMatchInlineSnapshot(`
     "------------------------------------------------------------------
                     🌲⚡️ Welcome to RedwoodJS! ⚡️🌲
     ------------------------------------------------------------------
     6.0.7
     [?25l[?25h"
   `)
  ```